Separating Client and Host
Changes created are meant to separate the code to allow the client and host.
Initially all the code was bundled together by react by having the client built and then sent by the host when connecting to it to act as a monolithic service.
The purpose of separating them was to permit setting up the project as a bundle of micro services making it easier to develop and manage.
This also allows any future changes to a framework or different service all together easier as well as deployment doesn't have to be on the same machine/docker container.
ENV Examples were updated accordingly to better represent the code.