Dockerization of Application
Initial docker setup of the project for both client and host sides.
Individual ports set for now have requirement setups within their respective .ENV files.
Redis is required as well but will be set within a separate Docker-Compose.
Important note is that individual volume setups are currently not in place based initially from a deployment perspective and is an important change that needs to be considered.