Deployment Setup
The current setup consists of using a docker compose to launch the three required containers for the project.
Initial Terraform setup was left to its own configuration using Ansible for the provisioning.
Nearly all code was moved out from the config due to long redployment times and setups on changes which could possibly break the cloud machine.
The post launch script only remains for the IP tables, this will eventually be moved into the ansible config as well with their built in model and has been left currently due to a bug on deployment possibly related to the IP table route.
It should be noted that while the memory was gradually increased due to performance issues, Docker crashing is still something that needs to be considered as Entropy remains relatively low at 256.
Previously this was a bigger issue, build times for the server currently sit at around ~15 seconds, while the client at ~200 seconds.