Development Machine Setup
Currently set for those who do not have Linux installed and wish to use a VM for the project, this setup is meant to be used as such.
At the current moment it was setup primarily to run any applications meant for Linux (ex: Ansible) without requiring too much to be installed locally.
Bash script may be used on your own machine if desired.
Future considerations may move this to Python if necessary.