Launch Script Setup
In order to automate the process and massively reduce times between manual enteries and verifications an initial draft of a script that could be used for setups has been created.
Much of it currently needs to be refactored however launch, teardowns, status and help has been created and is sufficient for now.
Future considerations besides massive refactoring include:
- a development setup of ENV variables
- logging instead of the required use of an active terminal
- Socket Replacement for SSH connection to the deployment server