Updated Dependencies
This includes modifying to use the serve command now as well as a general update of the package-lock's files.
Typically this would be left as is for the developers going forward however there were some issues in implementation.
Serve doesn't use the port automatically as React does which leaves it to either the developer to use the variable within the .env file or for future work to start moving into setting this via script or from the Docker configs.