Local Virtual Machine now Capable of Development
Previously the VM was set only with the necessary settings to be able to launch the deployment.
Now the VM is capable of launching the deployment server but also of being able to run the application on itself.
Very useful for others to use as reference as well if a VM isn't desired for the possible dependencies needed.
Ram was increased on the VM improving Docker compose times showing the issue on launching doesn't have to do with haveged.
Changes within the Docker compose in comparison to the deployment showcase no longer the need for a hardcoded IP for redis.
Repository setup was moved to the end of the provision script for future placement outside of it to setup updates without needing initialization or teardowns.