Launchscript Overhaul
This is the beginning of the overhaul.
Launch script was refactored heavily to now include only the arg parsers and general overlay of obtaining commands from the commands directory.
Provided the commands follow the ScriptCommand class like the others, simply placing it within the folder is enough to add it.
As commands grew and new functionality, it became difficult as modifications cause possible breaks in other commands as well as the file size kept growing.
Now only cases for both dveelopment and deployment servers are in common and can be imported as needed.