Jenkins Update
This commit brings in the necessary changes initially included for attempting to setup a Jenkins server.
This will need to be revised as Jenkins will be dumped in favor of GitHub Actions.
Currently in place for possibility of revert, launchScript errors and possible revision depending on the outcome of the setup tbd.