Updated Update Command
The update command has now been cleaned up as well as the main provisioning script as well to now allow connections to update the target branch.
It is no longer needed to manually have to change the code or log in to have this occur allowing easier testing as well.
Docker Relaunch setup was also added however not fully functionable due to the same issue with launching with synchronous terminal logging.