Matching Launch Command Behavior
By Attempting to remove the exit due to currently being unstable as a fully working solution has not yet been created, both Launch commands will not attempt to have the same behavior by launching the Application in both instances.
In the case of termination simply restarting the terminal will be necessary.